Error Handling Refactor
After reviewing with mentor the error handling path i chose to take, using an array of errors and passing them, we decided it was too much, and also too many validation layers
Now user input validation happens in one place, "trySubmit" which is ran for both adding and editing customers.

Created Date extension for formatting
removed equals for edit
removed Result Void SaveCustomerValidation error setup to simplify error handling
This removes multiple error checks throughout both customer VM's and the saveCustomer Interactor
Added Field Validation using is empty and a ternary based on the result
Removed validateLiveAndUpdateErrors to simplify error handling
Validation is performed in the trySubmit function
Modified editVM and addVM to return their useCase VM's
passed the SaveCustomer instance in the ListVM so its created with the VM and not in the app open.