Beginning of building the QuoteView
QuoteModel
Added: Any missing required fields to allow conversion of the child model representing a industry type, to the parent QuoteModel
Added: Function for converting the model, non optional value, all child quote models must have it to convert their model into a QuoteModel for saving
Added: CustomField struct to take in label and value these are for visual representation with in a quote that is saved, to show the values of industry specific fields.
Removed: Quoteable Conformance from the QuoteModel due to adding a conversion function for the other models to hold their own conversions. (toQuoteModel)
Updated: customField property to take in a struct made to pass in the values of industry specific fields.
Added: enum QuoteFields to monitor fields in the view that are active or inactive based on Industry Type.

CustomerListVM:
Added: closure newCustomerFromQuote to take in CustomerModel and hold it so when a customer is added through the quote view they can be chosen to assign the quote to.
Added: second addVM function that allows access to the onSubmit closure so we are able to add a customer from another view, and add a few more specific conditions to take that newly added customer and apply it to the current quote.

MainTabView:
Added: Switch for newValue in the onChange closure, now we are able to access the AddCustomerView through the QuoteView using the enum case .addCustomerToQuote, this is the utilization of the second addVM which allows access to the on submit closure, this allows us to call the addCustomer function, store the customer created after addition, then clear the value after application, this is all activated by the change in the enum.
added the case to the .sheet to pass the same data for the view to work with.

ActiveUserSheet:
Added: Case addCustomerFromQuote to allow another way to reach the AddCustomerView(sheet) with different expectations when closing the view after customer addition.